+++ apps/coordinator/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.11.1
+
+### Patch Changes
+
+- [#355](https://github.com/caravan-bitcoin/caravan/pull/355) [`ed62d7c`](https://github.com/caravan-bitcoin/caravan/commit/ed62d7cc7e2e731a590429c9e092b86a21d5ea1e) Thanks [@bucko13](https://github.com/bucko13)! - add input sequence support for caravan coordinator, psbt transformations, and trezor. Coordinator also gets some cleanup of old patterns that can be circumvented with better psbt tooling
+
+- Updated dependencies [[`ed62d7c`](https://github.com/caravan-bitcoin/caravan/commit/ed62d7cc7e2e731a590429c9e092b86a21d5ea1e), [`ed62d7c`](https://github.com/caravan-bitcoin/caravan/commit/ed62d7cc7e2e731a590429c9e092b86a21d5ea1e)]:
+  - @caravan/wallets@0.7.1
+  - @caravan/psbt@2.0.2
+
 ## 1.11.0
 
 ### Minor Changes